feat: Enhance user session booking flow and data handling
This commit introduces several enhancements to the user session booking flow and data handling:

- **Enhanced Session Booking Flow:**
    - Added a "بازگشت" (Back) button to the session date view, allowing users to return to the session list.
    - Modified the `book_session` function to include the date string in the callback data, enabling navigation back to the selected date's sessions.
    - Improved the session confirmation message with Persian day names and session details.
    - Added a "بازگشت به پنل سانس ها" (Back to Session Panel) button to the confirmation message.
- **Data Type Adjustments:**
    - Changed the `session_cost` and `amount` columns in the `PaymentCategory` and `Payment` models, respectively, from `DECIMAL` to `Integer` for simplified cost management.
    - Updated the `cost` column in the `Session` model from `DECIMAL` to `Integer`.
- **Session Availability Message:**
    - Modified the session availability message to be displayed as an edited message in response to a callback query.
- **Bug Fixes:**
    - Fixed a bug where the `show_sessions` function would throw an error if called from a callback query.

+++ repositories/models.py
@@ -56,7 +56,7 @@ class PaymentCategory(Base):
     __tablename__ = "payment_categories"
     id = Column(Integer, primary_key=True, autoincrement=True)
     account_type = Column(Enum(UserType), nullable=False)
-    session_cost = Column(DECIMAL(10), nullable=False)
+    session_cost = Column(Integer, nullable=False)
 
 
 class Session(Base):
@@ -66,7 +66,7 @@ class Session(Base):
     time_slot = Column(String(20), nullable=False)
     available = Column(Boolean, default=True)
     booked_user_id = Column(Integer, ForeignKey("users.user_id"))
-    cost = Column(DECIMAL(10), nullable=False)
+    cost = Column(Integer, nullable=False)
     user = relationship(
         "User", back_populates="sessions", foreign_keys=[booked_user_id]
     )
@@ -81,7 +81,7 @@ class Payment(Base):
     user_id = Column(Integer, ForeignKey("users.user_id"), nullable=False)
     session_id = Column(Integer, ForeignKey("sessions.id"), nullable=False)
     payment_date = Column(DateTime, nullable=False)
-    amount = Column(DECIMAL(10), nullable=False)
+    amount = Column(Integer, nullable=False)
     verified = Column(Boolean, default=False)
     user = relationship("User", back_populates="payments", foreign_keys=[user_id])
     session = relationship(
